python/crew_agents/pyproject.toml
[project]
name = "crew-agents"
version = "0.1.0"
description = "CrewAI agents for Rivermarsh game development"
requires-python = ">=3.13"
dependencies = [
    "crewai[anthropic]>=1.5.0",
    "crewai-tools>=1.5.0",
    "litellm>=1.80.0",
    "pyyaml>=6.0.3",
    "httpx>=0.28.1",
    "psutil>=7.1.3",
    "pydantic>=2.12.4",
    "psycopg[binary]>=3.2.12",
    "mesh-toolkit",
    "pyngrok>=7.5.0",
    "pytest>=9.0.1",
    "pytest-mock>=3.15.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/crew_agents"]

[tool.uv.sources]
mesh-toolkit = { workspace = true, editable = true }
